+++ src/test/LogExceptionTraitTest.php
@@ -5,11 +5,11 @@ namespace WebArch\LogTools\Test;
 use Exception;
 use Monolog\Handler\TestHandler;
 use Monolog\Logger;
-use PHPUnit\Framework\TestCase;
 use Psr\Log\LogLevel;
 use WebArch\LogTools\Test\Fixture\FooService;
+use WebArch\LogTools\Test\Fixture\LogExceptionFixture;
 
-class LogExceptionTraitTest extends TestCase
+class LogExceptionTraitTest extends LogExceptionFixture
 {
     /**
      * @var FooService
@@ -20,11 +20,6 @@ class LogExceptionTraitTest extends TestCase
      * @var TestHandler
      */
     private $testHandler;
-
-    /**
-     * @var string
-     */
-    private $rootDir;
 
     /**
      * @inheritDoc
@@ -43,126 +38,232 @@ class LogExceptionTraitTest extends TestCase
         );
     }
 
-    public function testLogException()
+    public function testLogExceptionDefaultChaining()
     {
-        /**
-         * Don't move the triggerExceptionCascade() call, or the following expectations will be needed to change!
-         */
-        $this->fooService->triggerExceptionCascade();
+        $this->fooService->triggerExceptionCascade(null);
+
         $expectedRecords = [
             [
-                'message' => $this->replaceRoot(
-                    '[InvalidArgumentException] Some argument is invalid (1) in %ROOT%/log-tools/src/test/Fixture/FooService.php:36'
-                ),
+                'message' => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
                 'context' => [
                     'foo'      => 'bar',
-                    'trace'    => $this->replaceRoot(
-                        '#0 %ROOT%/log-tools/src/test/Fixture/FooService.php(20): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()
-#1 %ROOT%/log-tools/src/test/LogExceptionTraitTest.php(51): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade()
-#2 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(1415): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogException()
-#3 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(1035): PHPUnit\\Framework\\TestCase->runTest()
-#4 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(691): PHPUnit\\Framework\\TestCase->runBare()
-#5 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(763): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))
-#6 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(597): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))
-#7 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(597): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))
-#8 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(621): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))
-#9 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(204): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)
-#10 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(163): PHPUnit\\TextUI\\Command->run(Array, true)
-#11 %ROOT%/log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()
-#12 {main}'
-                    ),
-                    'previous' => $this->replaceRoot(
-                        '[RuntimeException] Something is not found (2) in %ROOT%/log-tools/src/test/Fixture/FooService.php:49'
-                    ),
+                    'trace'    =>
+                        [
+                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
+                            1  => '#1 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(NULL)',
+                            2  => '#2 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionDefaultChaining()',
+                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
+                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
+                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
+                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
+                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
+                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
+                            11 => '#11 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
+                            12 => '#12 {main}',
+                        ],
+                    'previous' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
                 ],
-                'level'   => 600,
+                'level_name'   => 'EMERGENCY',
             ],
             [
-                'message' => $this->replaceRoot(
-                    '[RuntimeException] Something is not found (2) in %ROOT%/log-tools/src/test/Fixture/FooService.php:49'
-                ),
+                'message' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
                 'context' => [
                     'foo'      => 'bar',
-                    'trace'    => $this->replaceRoot(
-                        '#0 %ROOT%/log-tools/src/test/Fixture/FooService.php(34): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()
-#1 %ROOT%/log-tools/src/test/Fixture/FooService.php(20): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()
-#2 %ROOT%/log-tools/src/test/LogExceptionTraitTest.php(51): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade()
-#3 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(1415): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogException()
-#4 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(1035): PHPUnit\\Framework\\TestCase->runTest()
-#5 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(691): PHPUnit\\Framework\\TestCase->runBare()
-#6 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(763): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))
-#7 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(597): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))
-#8 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(597): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))
-#9 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(621): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))
-#10 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(204): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)
-#11 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(163): PHPUnit\\TextUI\\Command->run(Array, true)
-#12 %ROOT%/log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()
-#13 {main}'
-                    ),
-                    'previous' => $this->replaceRoot(
-                        '[LogicException] The error in the code (3) in %ROOT%/log-tools/src/test/Fixture/FooService.php:59'
-                    ),
+                    'trace'    =>
+                        [
+                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()',
+                            1  => '#1 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
+                            2  => '#2 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(NULL)',
+                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionDefaultChaining()',
+                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
+                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
+                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
+                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
+                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
+                            11 => '#11 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
+                            12 => '#12 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
+                            13 => '#13 {main}',
+                        ],
+                    'previous' => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
                 ],
-                'level'   => 600,
+                'level_name'   => 'EMERGENCY',
             ],
             [
-                'message' => $this->replaceRoot(
-                    '[LogicException] The error in the code (3) in %ROOT%/log-tools/src/test/Fixture/FooService.php:59'
-                ),
+                'message' => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
                 'context' => [
                     'foo'   => 'bar',
-                    'trace' => $this->replaceRoot(
-                        '#0 %ROOT%/log-tools/src/test/Fixture/FooService.php(47): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerLogicException()
-#1 %ROOT%/log-tools/src/test/Fixture/FooService.php(34): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()
-#2 %ROOT%/log-tools/src/test/Fixture/FooService.php(20): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()
-#3 %ROOT%/log-tools/src/test/LogExceptionTraitTest.php(51): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade()
-#4 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(1415): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogException()
-#5 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(1035): PHPUnit\\Framework\\TestCase->runTest()
-#6 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(691): PHPUnit\\Framework\\TestCase->runBare()
-#7 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(763): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))
-#8 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(597): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))
-#9 %ROOT%/log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(597): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))
-#10 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(621): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))
-#11 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(204): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)
-#12 %ROOT%/log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(163): PHPUnit\\TextUI\\Command->run(Array, true)
-#13 %ROOT%/log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()
-#14 {main}'
-                    ),
+                    'trace' =>
+                        [
+                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerLogicException()',
+                            1  => '#1 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()',
+                            2  => '#2 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
+                            3  => '#3 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(NULL)',
+                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionDefaultChaining()',
+                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
+                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
+                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
+                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
+                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            11 => '#11 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
+                            12 => '#12 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
+                            13 => '#13 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
+                            14 => '#14 {main}',
+                        ],
                 ],
-                'level'   => 600,
+                'level_name'   => 'EMERGENCY',
             ],
 
         ];
+        $this->assertLogRecords($expectedRecords);
+    }
 
+    public function testLogExceptionChainingEnabled()
+    {
+        $this->fooService->triggerExceptionCascade(true);
+
+        $expectedRecords = [
+            [
+                'message' => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context' => [
+                    'foo'      => 'bar',
+                    'trace'    =>
+                        [
+                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
+                            1  => '#1 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(true)',
+                            2  => '#2 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionChainingEnabled()',
+                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
+                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
+                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
+                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
+                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
+                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
+                            11 => '#11 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
+                            12 => '#12 {main}',
+                        ],
+                    'previous' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
+                ],
+                'level_name'   => 'EMERGENCY',
+            ],
+            [
+                'message' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context' => [
+                    'foo'      => 'bar',
+                    'trace'    =>
+                        [
+                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()',
+                            1  => '#1 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
+                            2  => '#2 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(true)',
+                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionChainingEnabled()',
+                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
+                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
+                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
+                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
+                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
+                            11 => '#11 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
+                            12 => '#12 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
+                            13 => '#13 {main}',
+                        ],
+                    'previous' => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
+                ],
+                'level_name'   => 'EMERGENCY',
+            ],
+            [
+                'message' => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context' => [
+                    'foo'   => 'bar',
+                    'trace' =>
+                        [
+                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerLogicException()',
+                            1  => '#1 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()',
+                            2  => '#2 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
+                            3  => '#3 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(true)',
+                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionChainingEnabled()',
+                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
+                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
+                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
+                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
+                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            11 => '#11 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
+                            12 => '#12 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
+                            13 => '#13 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
+                            14 => '#14 {main}',
+                        ],
+                ],
+                'level_name'   => 'EMERGENCY',
+            ],
+
+        ];
+        $this->assertLogRecords($expectedRecords);
+    }
+
+    public function testLogExceptionChainingDisabled()
+    {
+        $this->fooService->triggerExceptionCascade(false);
+
+        $expectedRecords = [
+            [
+                'message' => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context' => [
+                    'foo'      => 'bar',
+                    'trace'    =>
+                        [
+                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
+                            1  => '#1 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(false)',
+                            2  => '#2 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionChainingDisabled()',
+                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
+                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
+                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
+                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
+                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
+                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
+                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
+                            11 => '#11 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
+                            12 => '#12 {main}',
+                        ],
+                    'previous' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
+                ],
+                'level_name'   => 'EMERGENCY',
+            ],
+        ];
+        $this->assertLogRecords($expectedRecords);
+    }
+
+    /**
+     * @param array $expectedRecords
+     *
+     * @return void
+     */
+    protected function assertLogRecords(array $expectedRecords): void
+    {
         foreach ($this->testHandler->getRecords() as $key => $record) {
             $expectedRecord = array_shift($expectedRecords);
             $this->assertEquals(
                 $expectedRecord['message'],
-                $record['message'],
+                $this->cleanUpString($record['message']),
                 'Equal message for record #' . $key
             );
             $this->assertEquals(
-                $expectedRecord['level'],
-                $record['level'],
+                $expectedRecord['level_name'],
+                $record['level_name'],
                 'Equal level for record #' . $key
             );
             $this->assertEqualsCanonicalizing(
                 $expectedRecord['context'],
-                $record['context'],
+                $this->cleanUpArray($record['context']),
                 'Equal context for record #' . $key
             );
         }
-
-        $this->assertTrue(true);
-    }
-
-    protected function replaceRoot(string $string): string
-    {
-        return str_replace(
-            '%ROOT%',
-            $this->rootDir,
-            $string
-        );
     }
 
 }