+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+1.3.0
+-----
+
+### Added:
+- Protected method `LogExceptionTrait::setChaining()` can be used to disable or enable exception chaining for the
+class, which uses `\WebArch\LogTools\Traits\LogExceptionTrait`, but not for single `LogExceptionTrait::logException()`
+call
+
+### Changed:
+- Chaining is enabled by default
+- Trace string in the context is replaced by the array of strings.
+
+### Removed:
+- Parameter `$enableChaining` is removed from `logException()` method
+
+
 1.2.0
 -----